use angular to inject changeNowService
change:

inject service instead of polluting window object
